requirements.txt
flask
Flask-PyMongo
nltk
dnspython
python-dotenv
google-cloud-vision
bs4
fastai
jupyterlab
opencv-python
pillow
